Changed main build script
git-svn-id: http://svn.osdrawer.net/photograbber@92 f582c0bb-8d2e-0410-8fae-e3a1772be7bf

+++ PG.sh
@@ -18,8 +18,10 @@ make clean
 cd ../plugins/ptp
 make clean
 cd ../..
-rm -r build
-
+BUILDDIR="./build"
+if [ -d "$BUILDDIR" ] ; then
+	rm -r build
+fi
 # 	Build the core
 make
 # 	Build the interface
